% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.R
\docType{data}
\name{SUBMISSION_DEADLINE}
\alias{SUBMISSION_DEADLINE}
\title{GSoC 2026 Submission Deadline}
\format{
An object of class \code{Date} of length 1.
}
\usage{
SUBMISSION_DEADLINE
}
\description{
GSoC 2026 Submission Deadline
}
\keyword{datasets}
