% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.R
\name{submit_proposal}
\alias{submit_proposal}
\title{Submit GSoC Proposal}
\usage{
submit_proposal()
}
\value{
\code{TRUE} if current date is on or before the deadline,
\code{FALSE} otherwise.
}
\description{
This function checks if you can still submit your GSoC proposal.
}
\details{
The submission deadline is fetched dynamically at install time:
\strong{\Sexpr[stage=install, results=text]{as.character(gsocprop:::gsoc_env$deadline)}}.

\emph{\Sexpr[stage=render, results=text]{gsocprop:::get_status_message()}}
}
\examples{
submit_proposal()
}
