% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.R
\name{submit_proposal}
\alias{submit_proposal}
\title{Submit GSoC Proposal}
\usage{
submit_proposal()
}
\value{
\code{TRUE} if current date is on or before the deadline,
\code{FALSE} otherwise.
}
\description{
Checks if you can still submit your GSoC proposal.
The deadline is stored in \code{gsoc_env$deadline}.
}
\examples{
submit_proposal()

}
