% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.R
\name{submit_proposal}
\alias{submit_proposal}
\title{Submit GSoC Proposal}
\usage{
submit_proposal()
}
\value{
\code{TRUE} if called before the deadline, \code{FALSE} otherwise.
}
\description{
Checks whether the current date is before the GSoC 2026
submission deadline. The deadline is \Sexpr{format(gsocprop::SUBMISSION_DEADLINE, "\%B \%d, \%Y")}.
}
\details{
Submit a GSoC Proposal
}
\examples{
submit_proposal()
}
